+++ firmware/sim/testbench/muon_charge_corr_test_v2_tb.vhd
@@ -1,0 +1,305 @@
+
+-- Description:
+-- Testbench for simulation of mass div DR
+
+-- Version history:
+-- HB 2020-08-25: first design
+
+library ieee;
+use ieee.std_logic_1164.all;
+use ieee.std_logic_arith.all;
+use ieee.std_logic_unsigned.all;
+use ieee.fixed_pkg.all;
+library UNISIM;
+use UNISIM.VCOMPONENTS.ALL;
+library std;                  -- for Printing
+use std.textio.all;
+
+use work.math_pkg.all;
+use work.gtl_pkg.all;
+
+entity muon_charge_corr_test_v2_tb is
+end muon_charge_corr_test_v2_tb;
+
+architecture beh of muon_charge_corr_test_v2_tb is
+
+    signal lhc_clk : std_logic;
+
+    constant LHC_CLK_PERIOD : time :=  25 ns;
+    
+    signal muon : muon_objects_array(0 to NR_MUON_OBJECTS-1) := (others => (others => '0'));
+    signal algo, algo_o : std_logic_vector(1 downto 0);
+
+-- Signal definition of pt, eta and phi for correlation conditions.
+    signal mu_bx_0_pt_vector: diff_inputs_array(0 to NR_MU_OBJECTS-1) := (others => (others => '0'));
+    signal mu_bx_0_upt_vector: diff_inputs_array(0 to NR_MU_OBJECTS-1) := (others => (others => '0'));
+    signal mu_bx_0_eta_integer: integer_array(0 to NR_MU_OBJECTS-1) := (others => 0);
+    signal mu_bx_0_phi_integer: integer_array(0 to NR_MU_OBJECTS-1) := (others => 0);
+    signal mu_bx_0_eta_integer_half_res: integer_array(0 to NR_MU_OBJECTS-1) := (others => 0);
+    signal mu_bx_0_phi_integer_half_res: integer_array(0 to NR_MU_OBJECTS-1) := (others => 0);
+    signal mu_bx_0_cos_phi: integer_array(0 to NR_MU_OBJECTS-1) := (others => 0);
+    signal mu_bx_0_sin_phi: integer_array(0 to NR_MU_OBJECTS-1) := (others => 0);
+
+-- Signal definition for cuts of correlation conditions.
+    signal mu_mu_bx_0_bx_0_deta_integer: dim2_max_eta_range_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => 0));
+    signal mu_mu_bx_0_bx_0_deta: deta_dphi_vector_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => (others => '0')));
+    signal mu_mu_bx_0_bx_0_dphi_integer: dim2_max_phi_range_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => 0));
+    signal mu_mu_bx_0_bx_0_dphi: deta_dphi_vector_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => (others => '0')));
+    signal mu_mu_bx_0_bx_0_deta_integer_half_res: dim2_max_eta_range_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => 0));
+    signal mu_mu_bx_0_bx_0_dphi_integer_half_res: dim2_max_phi_range_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => 0));
+    signal mu_mu_bx_0_bx_0_mass_inv_pt : mass_dim2_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1) := (others => (others => (others => '0')));
+
+-- Signal definition for muon charge correlations.
+    signal ls_charcorr_double_bx_0_bx_0, os_charcorr_double_bx_0_bx_0 : std_logic_2dim_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1);
+    signal ls_charcorr_triple_bx_0_bx_0, os_charcorr_triple_bx_0_bx_0 : std_logic_3dim_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1);
+    signal ls_charcorr_quad_bx_0_bx_0, os_charcorr_quad_bx_0_bx_0 : std_logic_4dim_array(0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1, 0 to NR_MU_OBJECTS-1);
+
+-- Signal definition for conditions names
+    signal invariant_mass3_i0 : std_logic;
+    signal invariant_mass3_i3 : std_logic;
+
+-- Signal definition for algorithms names
+    signal l1_triple_mu_2_sq_1p5_sq_0_oq_mass_max12 : std_logic;
+    signal l1_triple_mu_3_sq_2p5_sq_0_oq_os_mass_max12 : std_logic;
+
+--*********************************Main Body of Code**********************************
+begin
+
+    -- Clock
+    process
+    begin
+        lhc_clk  <=  '1';
+        wait for LHC_CLK_PERIOD/2;
+        lhc_clk  <=  '0';
+        wait for LHC_CLK_PERIOD/2;
+    end process;
+
+    process
+    begin
+        wait for 5 ns;
+        muon <= (others => X"0000000000000000");
+        wait for LHC_CLK_PERIOD;
+        muon <= (X"000b280f58e01d3d", X"000fd00f4a601dc6", X"000fa80f4a601dc1", others => X"0000000000000000");
+        wait for LHC_CLK_PERIOD;
+        muon <= (others => X"0000000000000000");
+        wait for LHC_CLK_PERIOD;
+        muon <= (X"0004400fc9205470", X"0003a80bcfe028a5", X"0003b00bcae0289e", others => X"0000000000000000");
+        wait for LHC_CLK_PERIOD;
+        muon <= (others => X"0000000000000000");
+        wait for LHC_CLK_PERIOD;
+        muon <= (X"000cc80f9ee03d8d", X"0005c00b596020d8", X"0002180b5fe01c5f", X"0006400b592024e4", others => X"0000000000000000");
+        wait for LHC_CLK_PERIOD;
+        muon <= (others => X"0000000000000000");
+        wait;
+    end process;
+
+ ------------------- Instantiate  modules  -----------------
+
+cond_invariant_mass3_i0_i: entity work.correlation_conditions
+    generic map(
+-- slices for muon
+        slice_low_obj1 => 0,
+        slice_high_obj1 => 7,
+        slice_low_obj2 => 0,
+        slice_high_obj2 => 7,
+        slice_low_obj3 => 0,
+        slice_high_obj3 => 7,
+-- obj cuts
+        pt_threshold_obj1 => X"0005",
+        qual_lut_obj1 => X"F000",
+        pt_threshold_obj2 => X"0004",
+        qual_lut_obj2 => X"F000",
+        pt_threshold_obj3 => X"0001",
+        qual_lut_obj3 => X"FFF0",
+-- correlation cuts
+        mass_upper_limit_vector => X"00000000044AA200",
+        mass_lower_limit_vector => X"0000000000000000",
+        mass_vector_width => MU_PT_VECTOR_WIDTH+MU_PT_VECTOR_WIDTH+MU_MU_COSH_COS_VECTOR_WIDTH,
+        mass_3_obj => true,
+-- number of objects and type
+        nr_obj1 => NR_MU_OBJECTS,
+        type_obj1 => MU_TYPE,
+        nr_obj2 => NR_MU_OBJECTS,
+        type_obj2 => MU_TYPE,
+        nr_obj3 => NR_MU_OBJECTS,
+        type_obj3 => MU_TYPE,
+-- selector same/different bunch crossings
+        same_bx => true
+    )
+    port map(
+        lhc_clk,
+        muon_obj1 => muon,
+        muon_obj2 => muon,
+        muon_obj3 => muon,
+        mass_inv_pt => mu_mu_bx_0_bx_0_mass_inv_pt,
+        condition_o => invariant_mass3_i0
+    );
+
+cond_invariant_mass3_i3_i: entity work.correlation_conditions
+    generic map(
+-- slices for muon
+        slice_low_obj1 => 0,
+        slice_high_obj1 => 7,
+        slice_low_obj2 => 0,
+        slice_high_obj2 => 7,
+        slice_low_obj3 => 0,
+        slice_high_obj3 => 7,
+-- obj cuts
+        pt_threshold_obj1 => X"0007",
+        qual_lut_obj1 => X"F000",
+        pt_threshold_obj2 => X"0006",
+        qual_lut_obj2 => X"F000",
+        pt_threshold_obj3 => X"0001",
+        qual_lut_obj3 => X"FFF0",
+-- correlation cuts
+        mass_upper_limit_vector => X"00000000044AA200",
+        mass_lower_limit_vector => X"0000000000000000",
+        mass_vector_width => MU_PT_VECTOR_WIDTH+MU_PT_VECTOR_WIDTH+MU_MU_COSH_COS_VECTOR_WIDTH,
+        mass_3_obj => true,
+-- number of objects and type
+        nr_obj1 => NR_MU_OBJECTS,
+        type_obj1 => MU_TYPE,
+        nr_obj2 => NR_MU_OBJECTS,
+        type_obj2 => MU_TYPE,
+        nr_obj3 => NR_MU_OBJECTS,
+        type_obj3 => MU_TYPE,
+-- requested charge correlation
+        requested_charge_correlation => "os",
+-- selector same/different bunch crossings
+        same_bx => true
+    )
+    port map(
+        lhc_clk,
+        muon_obj1 => muon,
+        muon_obj2 => muon,
+        muon_obj3 => muon,
+        ls_charcorr_triple => ls_charcorr_triple_bx_0_bx_0,
+        os_charcorr_triple => os_charcorr_triple_bx_0_bx_0,
+        mass_inv_pt => mu_mu_bx_0_bx_0_mass_inv_pt,
+        condition_o => invariant_mass3_i3
+    );
+
+-- External condition assignment
+
+
+-- ========================================================
+-- Instantiations of algorithms
+
+-- 0 L1_TripleMu_2SQ_1p5SQ_0OQ_Mass_Max12 : mass_inv_3{MU2[MU-QLTY_SNGL],MU1p5[MU-QLTY_SNGL],MU0[MU-QLTY_OPEN]}[MASS_MAX_12]
+l1_triple_mu_2_sq_1p5_sq_0_oq_mass_max12 <= invariant_mass3_i0;
+algo(0) <= l1_triple_mu_2_sq_1p5_sq_0_oq_mass_max12;
+
+-- 3 L1_TripleMu_3SQ_2p5SQ_0OQ_OS_Mass_Max12 : mass_inv_3{MU3[MU-QLTY_SNGL],MU2p5[MU-QLTY_SNGL],MU0[MU-QLTY_OPEN]}[CHGCOR_OS,MASS_MAX_12]
+l1_triple_mu_3_sq_2p5_sq_0_oq_os_mass_max12 <= invariant_mass3_i3;
+algo(1) <= l1_triple_mu_3_sq_2p5_sq_0_oq_os_mass_max12;
+
+-- ========================================================
+-- Instantiations conversions, calculations, etc.
+-- eta and phi conversion to muon scale for calo-muon and muon-esums correlation conditions (used for DETA, DPHI, DR and mass)
+
+-- pt, eta, phi, cosine phi and sine phi for correlation conditions (used for DETA, DPHI, DR, mass, overlap_remover and two-body pt)
+
+calc_obj_parameter_mu_bx_0_i: entity work.obj_parameter
+    generic map(
+        nr_obj => NR_MU_OBJECTS,
+        type_obj => MU_TYPE
+    )
+    port map(
+        muon => muon,
+        pt_vector => mu_bx_0_pt_vector,
+        upt_vector => mu_bx_0_upt_vector,
+        eta_integer => mu_bx_0_eta_integer,
+        phi_integer => mu_bx_0_phi_integer,
+        eta_integer_h_r => mu_bx_0_eta_integer_half_res,
+        phi_integer_h_r => mu_bx_0_phi_integer_half_res,
+        cos_phi => mu_bx_0_cos_phi,
+        sin_phi => mu_bx_0_sin_phi
+    );
+--
+-- deta and dphi calculations for correlation conditions (used for DETA, DPHI)
+
+calc_deta_dphi_integer_mu_mu_bx_0_bx_0_i: entity work.deta_dphi_calculations
+    generic map(
+        phi_half_range => MUON_PHI_HALF_RANGE_BINS,
+        phi_h_r_half_range => MUON_PHI_HALF_RES_HALF_RANGE_BINS,
+        nr_obj1 => NR_MU_OBJECTS,
+        type_obj1 => MU_TYPE,
+        nr_obj2 => NR_MU_OBJECTS,
+        type_obj2 => MU_TYPE
+    )
+    port map(
+        eta_integer_obj1 => mu_bx_0_eta_integer,
+        phi_integer_obj1 => mu_bx_0_phi_integer,
+        eta_integer_obj2 => mu_bx_0_eta_integer,
+        phi_integer_obj2 => mu_bx_0_phi_integer,
+        eta_integer_h_r_obj1 => mu_bx_0_eta_integer_half_res,
+        phi_integer_h_r_obj1 => mu_bx_0_phi_integer_half_res,
+        eta_integer_h_r_obj2 => mu_bx_0_eta_integer_half_res,
+        phi_integer_h_r_obj2 => mu_bx_0_phi_integer_half_res,
+        deta_integer_half_res => mu_mu_bx_0_bx_0_deta_integer_half_res,
+        dphi_integer_half_res => mu_mu_bx_0_bx_0_dphi_integer_half_res,
+        deta_integer => mu_mu_bx_0_bx_0_deta_integer,
+        dphi_integer => mu_mu_bx_0_bx_0_dphi_integer
+    );
+--
+-- eta, dphi, cosh deta and cos dphi LUTs for correlation conditions (used for DR and mass)
+--
+-- Instantiations of correlation cuts calculations
+--
+-- Instantiations of DeltaEta LUTs
+
+-- Instantiations of DeltaPhi LUTs
+
+-- Instantiations of DeltaR calculation
+
+-- Instantiations of Invariant mass calculation
+
+calc_cut_mass_inv_pt_mu_mu_bx_0_bx_0_i: entity work.correlation_cuts_calculation
+    generic map(
+        nr_obj1 => NR_MU_OBJECTS,
+        type_obj1 => MU_TYPE,
+        nr_obj2 => NR_MU_OBJECTS,
+        type_obj2 => MU_TYPE,
+        mass_cut => true,
+        mass_type => INVARIANT_MASS_TYPE,
+        pt1_width => MU_PT_VECTOR_WIDTH,
+        pt2_width => MU_PT_VECTOR_WIDTH,
+        cosh_cos_width => MUON_MUON_COSH_COS_VECTOR_WIDTH
+    )
+    port map(
+        deta_integer => mu_mu_bx_0_bx_0_deta_integer,
+        dphi_integer => mu_mu_bx_0_bx_0_dphi_integer,
+        pt1 => mu_bx_0_pt_vector,
+        pt2 => mu_bx_0_pt_vector,
+        inv_mass_pt => mu_mu_bx_0_bx_0_mass_inv_pt
+    );
+
+-- Instantiations of Invariant mass divided DeltaR calculation
+
+-- Instantiations of Invariant mass unconstrained pt calculation
+
+-- Instantiations of Transverse mass calculation
+
+-- Instantiations of Two-body pt calculation
+
+-- muon charge correlations
+
+calc_muon_charge_correlations_bx_0_bx_0_i: entity work.muon_charge_correlations
+    port map(muon, muon,
+        ls_charcorr_double_bx_0_bx_0, os_charcorr_double_bx_0_bx_0,
+        ls_charcorr_triple_bx_0_bx_0, os_charcorr_triple_bx_0_bx_0,
+        ls_charcorr_quad_bx_0_bx_0, os_charcorr_quad_bx_0_bx_0);
+--
+
+-- ========================================================
+
+
+-- One pipeline stages for algorithms
+algo_pipeline_p: process(lhc_clk, algo)
+    begin
+    if (lhc_clk'event and lhc_clk = '1') then
+        algo_o <= algo;
+    end if;
+end process;
+
+end beh;